Fix more icon
